refactor(frontend): extract review display to component
