+++ apps/frontend/src/lib/components/MediaItem.tsx
@@ -1,11 +1,14 @@
 import { APP_ROUTES } from "@/lib/constants";
-import { useCommitMedia } from "@/lib/hooks/graphql";
+import { useCommitMedia, useUser } from "@/lib/hooks/graphql";
 import { gqlClient } from "@/lib/services/api";
 import { Verb, getLot, getVerb } from "@/lib/utilities";
 import {
+	ActionIcon,
 	Anchor,
+	Avatar,
 	Box,
 	Button,
+	Collapse,
 	Flex,
 	Image,
 	Loader,
@@ -13,6 +16,7 @@ import {
 	Text,
 	Tooltip,
 } from "@mantine/core";
+import { useDisclosure } from "@mantine/hooks";
 import { notifications } from "@mantine/notifications";
 import {
 	AddMediaToCollectionDocument,
@@ -20,16 +24,97 @@ import {
 	type MediaSearchQuery,
 	MetadataLot,
 	MetadataSource,
+	type ReviewItem,
 } from "@ryot/generated/graphql/backend/graphql";
 import { changeCase, getInitials } from "@ryot/utilities";
-import { IconStarFilled } from "@tabler/icons-react";
+import { IconEdit, IconStarFilled } from "@tabler/icons-react";
 import { useMutation } from "@tanstack/react-query";
+import { DateTime } from "luxon";
 import Link from "next/link";
 import { useRouter } from "next/router";
 import { withQuery } from "ufo";
 
-export const MediaScrollArea = (props: {	children: JSX.Element}) => {
+export const MediaScrollArea = (props: { children: JSX.Element }) => {
 	return <ScrollArea.Autosize mah={300}>{props.children}</ScrollArea.Autosize>;
+};
+
+export const ReviewItemDisplay = ({
+	review,
+	metadataId,
+	creatorId,
+}: {
+	review: ReviewItem;
+	metadataId?: number;
+	creatorId?: number;
+}) => {
+	const [opened, { toggle }] = useDisclosure(false);
+	const user = useUser();
+
+	return (
+		<Box key={review.id} data-review-id={review.id}>
+			<Flex align={"center"} gap={"sm"}>
+				<Avatar color="cyan" radius="xl">
+					{getInitials(review.postedBy.name)}{" "}
+				</Avatar>
+				<Box>
+					<Text>{review.postedBy.name}</Text>
+					<Text>{DateTime.fromJSDate(review.postedOn).toLocaleString()}</Text>
+				</Box>
+				{user && user.id === review.postedBy.id ? (
+					<Link
+						href={withQuery(APP_ROUTES.media.individualMediaItem.postReview, {
+							metadataId,
+							creatorId,
+							reviewId: review.id,
+						})}
+						passHref
+						legacyBehavior
+					>
+						<Anchor>
+							<ActionIcon>
+								<IconEdit size="1rem" />
+							</ActionIcon>
+						</Anchor>
+					</Link>
+				) : null}
+			</Flex>
+			<Box ml={"sm"} mt={"xs"}>
+				{typeof review.showSeason === "number" ? (
+					<Text color="dimmed">
+						S{review.showSeason}-E
+						{review.showEpisode}
+					</Text>
+				) : null}
+				{typeof review.podcastEpisode === "number" ? (
+					<Text color="dimmed">EP-{review.podcastEpisode}</Text>
+				) : null}
+				{review.rating > 0 ? (
+					<Flex align={"center"} gap={4}>
+						<IconStarFilled size={"1rem"} style={{ color: "#EBE600FF" }} />
+						<Text color="white" fw="bold">
+							{review.rating} %
+						</Text>
+					</Flex>
+				) : null}
+				{review.text ? (
+					!review.spoiler ? (
+						<Text dangerouslySetInnerHTML={{ __html: review.text }} />
+					) : (
+						<>
+							{!opened ? (
+								<Button onClick={toggle} variant={"subtle"} compact>
+									Show spoiler
+								</Button>
+							) : null}
+							<Collapse in={opened}>
+								<Text dangerouslySetInnerHTML={{ __html: review.text }} />
+							</Collapse>
+						</>
+					)
+				) : null}
+			</Box>
+		</Box>
+	);
 };
 
 export const BaseDisplayItem = (props: {
